@@ -19,9 +19,9 @@
     "leaflet": "1.3.1",
     "prop-types": "15.6.1",
     "ramda": "0.25.0",
-    "react": "16.4.0",
+    "react": "16.7.0-alpha.0",
     "react-apollo": "2.1.11",
-    "react-dom": "16.4.0",
+    "react-dom": "16.7.0-alpha.0",
     "react-leaflet": "1.8.2",
     "react-leaflet-control": "1.4.1",
     "react-router": "4.2.0",
