@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React, { useState } from 'react';
 import { Map } from 'react-leaflet';
 
 import {
@@ -16,113 +16,69 @@ import ParcelPolygon from './ParcelPolygon';
 
 const REACT_LEAFLET_MAP_ID = 'react-leaflet-map';
 
-export default class ReactLeafletMap extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            lat: null,
-            lng: null,
-            selectPointMode: false,
-            fetching: false,
-            queryMode: queryModeEnum.RWD,
-        };
+export default function ReactLeafletMap() {
+    const [selectPointMode, setSelectPointMode] = useState(false);
+    const [fetching, setFetching] = useState(false);
+    const [queryMode, setQueryMode] = useState(queryModeEnum.RWD);
+    const [lat, setLat] = useState(null);
+    const [lng, setLng] = useState(null);
 
-        this.handleMapClick = this.handleMapClick.bind(this);
-        this.toggleSelectPointMode = this.toggleSelectPointMode.bind(this);
-        this.stopFetching = this.stopFetching.bind(this);
-        this.toggleRWDParcelMode = this.toggleRWDParcelMode.bind(this);
-    }
+    const stopFetching = () => setFetching(false);
+    const toggleSelectPointMode = () => setSelectPointMode(!selectPointMode);
 
-    handleMapClick({ latlng }) {
-        return this.state.selectPointMode
-            ? this.setState(state => ({
-                ...state,
-                ...latlng,
-                selectPointMode: false,
-                fetching: true,
-            }))
-            : null;
-    }
+    const toggleRWDParcelMode = queryMode === queryModeEnum.RWD
+        ? () => setQueryMode(queryModeEnum.Parcel)
+        : () => setQueryMode(queryModeEnum.RWD);
 
-    toggleSelectPointMode() {
-        return this.setState(
-            state => ({
-                ...state,
-                selectPointMode: !state.selectPointMode,
-            }),
-            () => {
-                document
-                    .getElementById(REACT_LEAFLET_MAP_ID)
-                    .style
-                    .cursor = this.state.selectPointMode
-                        ? 'pointer'
-                        : null;
-            },
+    const handleMapClick = ({ latlng }) => {
+        if (!selectPointMode) {
+            return null;
+        }
+
+        setFetching(true);
+        setSelectPointMode(false);
+        setLat(latlng.lat);
+        setLng(latlng.lng);
+
+        return null;
+    };
+
+    const polygon = queryMode === queryModeEnum.RWD
+        ? (
+            <RWDPolygon
+                lat={lat}
+                lng={lng}
+                stopFetching={stopFetching}
+            />)
+        : (
+            <ParcelPolygon
+                lat={lat}
+                lng={lng}
+                stopFetching={stopFetching}
+            />
         );
-    }
 
-    toggleRWDParcelMode() {
-        return this.setState(state => ({
-            ...state,
-            queryMode: state.queryMode === queryModeEnum.RWD
-                ? queryModeEnum.Parcel
-                : queryModeEnum.RWD,
-        }));
-    }
-
-    stopFetching() {
-        return this.setState(state => ({
-            ...state,
-            fetching: false,
-        }));
-    }
-
-    render() {
-        const {
-            lat,
-            lng,
-            selectPointMode,
-            fetching,
-            queryMode,
-        } = this.state;
-
-        const polygon = queryMode === queryModeEnum.RWD
-            ? (
-                <RWDPolygon
-                    lat={lat}
-                    lng={lng}
-                    stopFetching={this.stopFetching}
-                />)
-            : (
-                <ParcelPolygon
-                    lat={lat}
-                    lng={lng}
-                    stopFetching={this.stopFetching}
-                />
-            );
-
-        return (
-            <Map
-                center={cityHallCoordinates}
-                zoom={initialMapZoom}
-                id={REACT_LEAFLET_MAP_ID}
-                onClick={this.handleMapClick}
-            >
-                <ReactLeafletEsriTiledMapLayer />
-                <ToggleRWDParcelModeControl
-                    loading={fetching}
-                    position={controlPositionsEnum.topleft}
-                    mode={queryMode}
-                    toggleRWDParcelMode={this.toggleRWDParcelMode}
-                />
-                <SelectPointControl
-                    loading={fetching}
-                    position={controlPositionsEnum.topleft}
-                    selectPointMode={selectPointMode}
-                    toggleSelectPointMode={this.toggleSelectPointMode}
-                />
-                {polygon}
-            </Map>
-        );
-    }
+    return (
+        <Map
+            center={cityHallCoordinates}
+            zoom={initialMapZoom}
+            id={REACT_LEAFLET_MAP_ID}
+            onClick={handleMapClick}
+        >
+            <ReactLeafletEsriTiledMapLayer />
+            <ToggleRWDParcelModeControl
+                loading={fetching}
+                position={controlPositionsEnum.topleft}
+                mode={queryMode}
+                toggleRWDParcelMode={toggleRWDParcelMode}
+            />
+            <SelectPointControl
+                loading={fetching}
+                position={controlPositionsEnum.topleft}
+                selectPointMode={selectPointMode}
+                toggleSelectPointMode={toggleSelectPointMode}
+            />
+            {polygon}
+        </Map>
+    );
 }
